requirements.txt
streamlit==1.15.0
farm-haystack==1.4.0
torch==1.12.0
transformers==4.25.1